Add support to scaffold Empty View  -Split the ViewGenerator into a commandline entrypoint and the actual  scaffolders, so we have a consistent pattern for controllers views and  dependency
